post_processor/requirements.txt
fastapi
motor
python-dotenv
uvicorn
pydantic
requests
